+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Supported enabling SMB Oauth on Storage accounts 
+    - `New-AzStorageAccount`
+    - `Set-AzStorageAccount`
+* Supported Enable Metrics when set object replication policy 
+    - `Set-AzStorageObjectReplicationPolicy`
 
 ## Version 9.1.0
 * Supported set SasExpirationAction as Log or Block, together with SasExpirationPeriod